抽取公共定义
